% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPDP_caret.R
\name{bsPDP.caret}
\alias{bsPDP.caret}
\title{Bootstrapped partial dependence plots for \code{caret} functions}
\usage{
bsPDP.caret(
  formula,
  variable = NULL,
  data,
  newdata = NULL,
  grid = NULL,
  outcome = NULL,
  n_boot = 100,
  p_boot = 0.6,
  N = 1000,
  label = NULL,
  clock = FALSE,
  test = FALSE,
  seed = 8675309,
  trainMethod = "rf",
  ...
)
}
\arguments{
\item{formula}{is an output formula for the outcome.}

\item{variable}{is the treatment variable.}

\item{data}{is a data frame to be used for the training.}

\item{newdata}{is an optional data frame of test data for the PDPs.}

\item{grid}{sets the values of \code{variable} to evaluate the default is 100 values in range of \code{variable}:
grid = seq(min(x), max(x), length.out = 100).}

\item{n_boot}{is the number of bootstrap replications.}

\item{p_boot}{is the proportion of the data to select for each bootstrap replication.}

\item{N}{is the number of observations to select for calculating the PDPs.}

\item{label}{is a character-string variable label for \code{variable}.}

\item{clock}{is a logical indicating whether to time each bootstrap replication.}

\item{test}{is a logical indicating whether to calculate the pdp for both the \code{data} and \code{newdata}.}

\item{seed}{is a random seed (default is 8675309).}

\item{trainMethod}{is the train method for the estimation default is \code{"rf"} (random forest). For details, see http://topepo.github.io/caret/using-your-own-model-in-train.html.}

\item{...}{additional arguments for the \code{train} function, see \code{?train} or http://topepo.github.io/caret/using-your-own-model-in-train.html. We recommended checking settings for \code{trControl = trainControl()} (control the computational nuances of the train function, see ?trainControl) and \code{tuneGrid} (a data frame with possible tuning values).}
}
\value{
\code{bsPDPcaret} returns an object with class "bsPDP," a list that includes the following components:
\item{variable}{the treatment variable.}
\item{pdpData}{the estimated average predictions and standard errors along \code{variable}.}
\item{trainData}{the original training data.}
\item{testData}{the test data.}
\item{outcome}{the outcome class (\code{NULL} for outcomes with \code{numeric} class).}
\item{method}{the method (\code{c("caret", method)})}
}
\description{
Evaluates a models in the \code{caret} package for \code{n_boot} bootstrap iterations to obtain confidence bands for the estimated PDP.
}
\references{
Yakusheva, Olga; Bang, James T; Bobay, Kathleen; Hughes, Ronda G; Costa, Linda; and Weiss, Marianne (2021, Forthcoming). \emph{Health Services Research.}
}
