% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPDP_hi_est.R
\name{bsPDP.hi_est}
\alias{bsPDP.hi_est}
\title{Bootstrapped partial dependence plots for the \code{hi_est} estimator}
\usage{
bsPDP.hi_est(
  formula,
  variable = NULL,
  data,
  newdata = NULL,
  grid = NULL,
  outcome = NULL,
  n_boot = 100,
  p_boot = 0.6,
  N = 1000,
  label = NULL,
  clock = FALSE,
  test = FALSE,
  seed = 8675309,
  treatment_formula = NULL,
  treatment_mod = NULL,
  link = NULL,
  ...
)
}
\arguments{
\item{formula}{is the outcome formula.}

\item{variable}{is the treatment variable.}

\item{data}{is a data frame to be used for the training.}

\item{newdata}{is an optional data frame of test data for the PDPs.}

\item{grid}{sets the values of \code{variable} to evaluate the default is 100 values in range of \code{variable}:
grid = seq(min(x), max(x), length.out = 100).}

\item{outcome}{is the outcome class to be predicted for classification problems.}

\item{n_boot}{is the number of bootstrap replications.}

\item{p_boot}{is the proportion of the data to select for each bootstrap replication.}

\item{N}{is the number of observations to select for calculating the PDPs.}

\item{label}{is a character-string variable label for \code{variable}.}

\item{clock}{is a logical indicating whether to time each bootstrap replication.}

\item{test}{is a logical indicating whether to calculate the pdp for both the \code{data} and \code{newdata}.}

\item{seed}{is a random seed (default is 8675309).}

\item{treatment_formula}{is a formula for covariate-balancing the treatment.}

\item{treatment_mod}{is the type of model for the treatment (default is "Binomial" for factor outcomes or "Normal" otherwise).}

\item{link}{indicates the type of link function (default is "logit" for \code{treatment_mod = "Binomial"} or \code{treatment_mod = "Ordinal"}).}

\item{...}{additional arguments for the \code{hi_est} function.}
}
\value{
\code{bsPDPglm} returns an object with class "bsPDP," a list that includes the following components:
\item{variable}{the treatment variable.}
\item{pdpData}{the estimated average predictions and standard errors along \code{variable}.}
\item{trainData}{the original training data.}
\item{testData}{the test data.}
\item{outcome}{the outcome class (\code{NULL} for outcomes with \code{numeric} class).}
\item{trControl}{is not applicable for this method (\code{NULL}).}
}
\description{
Evaluates a Hirano and Imbens (2004) estimator for \code{n_boot} bootstrap iterations to obtain confidence bands for the estimated PDP.
}
