% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPDP_randomForest.R
\name{bsPDP.randomForest}
\alias{bsPDP.randomForest}
\title{Bootstrapped partial dependence plots for \code{randomForest}}
\usage{
bsPDP.randomForest(
  formula,
  variable = NULL,
  data,
  newdata = NULL,
  grid = NULL,
  outcome = NULL,
  n_boot = 100,
  p_boot = 0.6,
  N = 1000,
  label = NULL,
  clock = FALSE,
  test = FALSE,
  seed = 8675309,
  parallel = TRUE,
  n_tree = 500,
  ...
)
}
\arguments{
\item{formula}{is an output formula for the outcome.}

\item{variable}{is the treatment variable.}

\item{data}{is a data frame to be used for the training.}

\item{newdata}{is an optional data frame of test data for the PDPs.}

\item{grid}{sets the values of \code{variable} to evaluate (default is 100 values in range of \code{variable}: \code{grid = seq(min(x), max(x), length.out = 100}).}

\item{outcome}{is the outcome class to be predicted for classification problems.}

\item{n_boot}{is the number of bootstrap replications.}

\item{p_boot}{is the proportion of the data to select for each bootstrap replication.}

\item{N}{is the number of observations to select for calculating the PDPs.}

\item{label}{is a character-string variable label for \code{variable}.}

\item{clock}{is a logical indicating whether to time each bootstrap replication.}

\item{test}{is a logical indicating whether to calculate the pdp for both the \code{data} and \code{newdata}.}

\item{seed}{is a random seed (default is 8675309).}

\item{parallel}{indicates whether to run the random forest algorithm in parallel (default = \code{TRUE}).}

\item{...}{additional arguments for \code{randomForest} function.}

\item{method}{indicates the type of model (randomForest).}
}
\value{
\code{bsPDPrf} returns an object with class "bsPDP," a list that includes the following components:
\item{variable}{the treatment variable.}
\item{pdpData}{the estimated average predictions and standard errors along \code{variable}.}
\item{trainData}{the original training data.}
\item{testData}{the test data.}
\item{outcome}{the outcome class (\code{NULL} for outcomes with \code{numeric} class).}
\item{trControl}{is not applicable for this method (\code{NULL}).}
\item{method}{the method (\code{BayesTree})}
}
\description{
Evaluates a \code{randomForest} model for \code{n_boot} bootstrap iterations to obtain confidence bands for the estimated PDP.
}
