% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPDP.R
\name{bsPDP}
\alias{bsPDP}
\title{Bootstrapped partial dependence plots}
\usage{
bsPDP(
  formula,
  variable = NULL,
  data,
  newdata = NULL,
  grid = NULL,
  outcome = NULL,
  n_boot = 100,
  p_boot = 0.6,
  N = 1000,
  method = "randomForest",
  label = NULL,
  clock = FALSE,
  test = FALSE,
  seed = 8675309,
  ...
)
}
\arguments{
\item{formula}{(required) an output formula for the outcome.}

\item{variable}{(required) the treatment variable.}

\item{data}{(required) a data frame to be used for the training.}

\item{newdata}{an optional data frame of test data for the PDPs (default is \code{data}).}

\item{grid}{sets the values of \code{variable} to evaluate (the default is 100 values in range of \code{variable}).}

\item{n_boot}{the number of bootstrap replications (default is 100).}

\item{p_boot}{the proportion of the data to select for each bootstrap replication (default is 0.6).}

\item{N}{the number of observations to select for calculating the PDPs (default is 1000).}

\item{method}{the model method (default is \code{'randomForest'}.}

\item{label}{a character-string variable label for \code{variable} (default is \code{variable}).}

\item{clock}{a logical indicating whether to time each bootstrap replication (default is \code{FALSE}.}

\item{test}{a logical indicating whether to calculate the pdp for both the \code{data} and \code{newdata} (default is \code{FALSE}.}

\item{seed}{a random seed (default is 8675309).}

\item{...}{additional arguments specific to the method some additional arguments may be required for some methods, e.g. \code{caret}, \code{iptw_est}, and \code{hi_est}.}

\item{plot}{a logical indicating whether to plot the resulting \code{bsPDP} object (default is \code{TRUE}.}
}
\value{
\code{bsPDP} returns an object with class "bsPDP," a list that includes the following components:
\item{variable}{the treatment variable.}
\item{pdpData}{the estimated average predictions and standard errors along \code{variable}.}
\item{trainData}{the original training data.}
\item{testData}{the test data.}
\item{outcome}{the outcome class (\code{NULL} for outcomes with \code{numeric} class).}
\item{method}{the method}
}
\description{
Evaluates models for \code{n_boot} bootstrap iterations to obtain confidence bands for the estimated PDP.
}
\details{
\code{bsPDP} currently supports the following methods:

\code{lm} for linear models

\code{glm} for generalized linear models

\code{randomForest} for random forests

\code{caret} for pramater tuning with machine-learning models (including tree ensembles and neural networks)

\code{dbarts} and \code{BayesTree} for Bayesian Additive Regression Trees (BART)

\code{causaldrf} for inverse probability-of-treatment weights estimators and the Hirano-Imbens (2004) covariate balancing estimators.
}
\references{
Yakusheva, Olga; Bang, James T; Bobay, Kathleen; Hughes, Ronda G; Costa, Linda; and Weiss, Marianne (2021, Forthcoming). \emph{Health Services Research.}

Breiman, Leo, 2001. Random forests. \emph{Machine learning.} 45(1), pp.5-32.

Chipman, H.; George, E.; and McCulloch, R. (2010) BART: Bayesian Additive Regression Trees. \emph{Annals of Statistics.} 4(1), pp. 266-298.

Hirano, Keisuke, Imbens, Guido W (2004).  The propensity score with continuous treatments.  \emph{Applied Bayesian modeling and causal inference from incomplete-data perspectives.} 226164, pp. 73-84.

Schafer, J.L., Galagate, D.L. (2015).  Causal inference with a continuous treatment and outcome: alternative estimators for parametric dose-response models. \emph{Doctoral dissertation}.
}
