% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_est.R
\name{hi_est}
\alias{hi_est}
\title{The Hirano and Imbens estimator}
\usage{
hi_est(
  Y,
  treat,
  treat_formula,
  outcome_formula,
  data,
  grid_val,
  treat_mod,
  link_function,
  ...
)
}
\arguments{
\item{Y}{is the the name of the outcome variable contained in \code{data}.}

\item{treat}{is the name of the treatment variable contained in
\code{data}.}

\item{treat_formula}{an object of class "formula" (or one that can be
coerced to that class) that regresses \code{treat} on a linear combination
of \code{X}: a symbolic description of the model to be fitted.}

\item{outcome_formula}{is the formula used for fitting the outcome surface. \code{gps} is one of the independent variables to use in the outcome_formula. ie.

\code{Y ~ treat+ I(treat^2) + gps + I(gps^2) + treat * gps}

or a variation of this.  Use \code{gps} as the name of the variable representing the gps in \code{outcome_formula}.}

\item{data}{is a dataframe containing \code{Y}, \code{treat}, and
\code{X}.}

\item{grid_val}{contains the treatment values to be evaluated.}

\item{treat_mod}{a description of the error distribution to be used in the
model for treatment. Options include: \code{"Normal"} for normal model,
\code{"LogNormal"} for lognormal model, \code{"Sqrt"} for square-root transformation
to a normal treatment, \code{"Poisson"} for Poisson model,
\code{"NegBinom"} for negative binomial model, \code{"Gamma"} for gamma
model, \code{"Binomial"} for binomial model.}

\item{link_function}{For \code{treat_mod = "Gamma"} (fitted using glm) alternatives are "log" or "inverse".
For \code{treat_mod = "Binomial"} (fitted using glm) alternatives are "logit", "probit", "cauchit", "log" and "cloglog".}

\item{...}{additional arguments to be passed to the outcome lm() function.}
}
\value{
\code{hi_est} returns an object of class "causaldrf",
a list that contains the following components:
\item{param}{parameter estimates for a hi fit.}
\item{t_mod}{the result of the treatment model fit.}
\item{out_mod}{the result of the outcome model fit.}
\item{call}{the matched call.}

##' @references Schafer, J.L., Galagate, D.L. (2015).  Causal inference with a
continuous treatment and outcome: alternative estimators for parametric
dose-response models. \emph{Manuscript in preparation}.
}
\description{
This function estimates the GPS function and estimates the
ADRF. The GPS score is based on different treatment models.  The treatment is
linearly related to Xs.
}
