% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_est.R
\name{iptw_est}
\alias{iptw_est}
\title{The iptw method or importance weighting method estimates the ADRF by
weighting the data with stabilized or non-stabilized weights.}
\usage{
iptw_est(
  Y,
  treat,
  treat_formula,
  numerator_formula,
  data,
  degree,
  treat_mod,
  link_function,
  ...
)
}
\arguments{
\item{Y}{is the the name of the outcome variable contained in \code{data}.}

\item{treat}{is the name of the treatment variable contained in
\code{data}.}

\item{treat_formula}{an object of class "formula" (or one that can be
coerced to that class) that regresses \code{treat} on a linear combination
of \code{X}: a symbolic description of the model to be fitted.}

\item{numerator_formula}{an object of class "formula" (or one that can be
coerced to that class) that regresses \code{treat} on a linear combination
of \code{X}: a symbolic description of the model to be fitted.  i.e.
\code{treat ~ 1}.}

\item{data}{is a dataframe containing \code{Y}, \code{treat}, and
\code{X}.}

\item{degree}{is 1 for linear and 2 for quadratic outcome model.}

\item{treat_mod}{a description of the error distribution to be used in the
model for treatment. Options include: \code{"Normal"} for normal model,
\code{"LogNormal"} for lognormal model, \code{"Sqrt"} for square-root transformation
to a normal treatment, \code{"Poisson"} for Poisson model,
\code{"NegBinom"} for negative binomial model, \code{"Gamma"} for gamma
model, \code{"Binomial"} for binomial model, \code{"Ordinal"} for ordinal model,
\code{"Multinomial"} for multinomial model.}

\item{link_function}{specifies the link function between the variables in
numerator or denominator and exposure, respectively.
For \code{treat_mod = "Gamma"} (fitted using glm) alternatives are "log" or "inverse".
For \code{treat_mod = "Binomial"} (fitted using glm) alternatives are "logit", "probit", "cauchit", "log" and "cloglog".
For \code{treat_mod = "Multinomial"} this argument is ignored, and
multinomial logistic regression models are always used (fitted using multinom).
For \code{treat_mod = "Ordinal"} (fitted using polr) alternatives are "logit", "probit", "cauchit", and "cloglog".}

\item{...}{additional arguments to be passed to the low level treatment regression fitting functions.}
}
\value{
\code{iptw_est} returns an object of class "causaldrf", a list that contains the following components:
\item{param}{parameter estimates for a iptw fit.}
\item{t_mod}{the result of the treatment model fit.}
\item{num_mod}{the result of the numerator model fit.}
\item{weights}{the estimated weights.}
\item{weight_data}{the weights.}
\item{out_mod}{the outcome model.}
}
\description{
The iptw method or importance weighting method estimates the ADRF by
weighting the data with stabilized or non-stabilized weights.
}
\references{
Schafer, J.L., Galagate, D.L. (2015).  Causal inference with a
continuous treatment and outcome: alternative estimators for parametric
dose-response models. \emph{Manuscript in preparation}.
}
