% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bsPDP.R
\name{plot.bsPDP}
\alias{plot.bsPDP}
\title{Plot bootstrapped partial dependence plots}
\usage{
\method{plot}{bsPDP}(
  bsPDP,
  test = FALSE,
  title = "Partial Dependence Plot",
  xlim = NULL,
  ylim = NULL,
  xtitle = NULL,
  ytitle = "Predicted Outcome",
  iqr = FALSE,
  hist = FALSE
)
}
\arguments{
\item{bsPDP}{(required) a list created by a \code{bsPDP} function and contains the following elements:
\enumerate{
\item \code{trainData}: a data frame containing the original training sample.
\item \code{variable}: the x-variable from \code{trainData} to predict and plot.
\item \code{pdpData}: data frame containing the following variables:
a. \code{"variable"}: grid values of the x-variable;
b. \code{pdpHat}: the average of the bootstrapped predictions at each grid value in the training sample;
c. \code{pdpSd}: the standard errors of the bootstrapped predictions at each grid value in the training sample;
d. \code{pdpHat.test}: the average of the bootstrapped predictions at each grid value in the test sample;
e. \code{pdpSd.test}: the standard errors of the bootstrapped predictions at each grid in the test sample;
\item \code{testData}: a data frame containing the test sample.
\item \code{outcome}: the outcome class (for classification).
\item \code{trControl}: the \code{trainControl} settings (for tuned \code{caret} methods).
\item \code{trainMethod}: the \code{method} used in the \code{caret} functionl.
}}

\item{test}{is a logical indicating whether to plot the pdp for the test data instead of the train data (default is \code{FALSE}).}

\item{title}{is a string for the main title of the plot (the default is "Partial Dependence Plot").}

\item{xlim}{is a vector combination equal to the minimum and maximum values of the x-axis (the default is the minimum and maximum values of the x-variable).}

\item{ylim}{is a vector combination equal to the minimum and maximum values of the y-axis (the default is the minimum of \code{pdpHat} minus 3 times the maximum value of \code{pdpSd} and the maximum of pdpHat plus 3 times the maximum value of \code{pdpSd}).}

\item{xtitle}{is a character string for the x-axis title (the default is the name of the x-variable)}

\item{ytitle}{is a character string for the y-axis title (the default is "Predicted Outcome")}

\item{iqr}{is a logical indicating whether to include lines at the IQR bounds of x (the default is \code{FALSE})}

\item{hist}{is a logical indicating whether to include a histogram of x (the default is \code{FALSE})}
}
\value{
\code{plot.bsPDP} returns a \code{ggplot} that can bee manipulated according to \code{ggplot} syntax.
}
\description{
Plots objects of class \code{bsPDP} created by bootstrapping a predictive or causal model to show confidence bands for the average predicted partial dependence plot of a variable.
}
\details{
\code{bsPDP} and \code{plot.bsPDP} currently support the following methods:

\code{lm} for linear models

\code{glm} for generalized linear models

\code{randomForest} for random forests

\code{caret} for pramater tuning with machine-learning models (including tree ensembles and neural networks)

\code{dbarts} and \code{BayesTree} for Bayesian Additive Regression Trees (BART)

\code{causaldrf} for inverse probability-of-treatment weights estimators \code{iptw_est} and the Hirano-Imbens (2004) covariate balancing estimators \code{hi_est}.
}
\references{
Yakusheva, Olga; Bang, James T; Bobay, Kathleen; Hughes, Ronda G; Costa, Linda; and Weiss, Marianne (2021, Forthcoming). \emph{Health Services Research.}
}
